requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.0.3 # Added for configuration management

# AI/ML
vosk==0.3.45
moviepy==1.0.3
unstructured==0.11.0
opencv-python==4.8.1.78
PyMuPDF==1.23.8
Pillow==10.1.0
google-generativeai==0.3.0
requests==2.31.0
langchain-postgres==0.0.4

# OCR (for document_processor_ocr.py)
pytesseract==0.3.10
pdf2image==1.16.3

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0